requirement.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.0.0
Werkzeug==0.10.4
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
gevent==1.0.2
greenlet==0.4.9
gunicorn==19.3.0
html5lib==0.999
itsdangerous==0.24
requests==2.8.1
six==1.5.2
urllib3==1.7.1
wheel==0.24.0
wsgiref==0.1.2